fix get current location,and move camera on map

+++ lib/models/placePredictions.dart
@@ -13,8 +13,8 @@ class PlacePredictions {
     id =json["id"];
     text =json["text"];
     place_name =json["place_name"];
-    lat = json["center"][1];
-    lng = json["center"][0];
+    lat = json["center"][1] as double;
+    lng = json["center"][0] as double;
 
   }
